requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
scipy>=1.7.0

# Time series analysis
statsmodels>=0.13.0
prophet>=1.1.0
pmdarima>=2.0.0

# Machine learning
scikit-learn>=1.0.0
torch>=1.12.0
torchvision>=0.13.0

# Visualization
plotly>=5.0.0
seaborn>=0.11.0

# Web interface
streamlit>=1.20.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Optional dependencies for advanced features
# Uncomment if you need these features:
# tensorflow>=2.8.0
# keras>=2.8.0
# xgboost>=1.6.0
# lightgbm>=3.3.0
# catboost>=1.0.0
# darts>=0.20.0
# sktime>=0.12.0
# tslearn>=0.5.0
